use Map<Long, Beer>

+++ app/src/main/java/com/example/beerapp/data/beer/BeersRepository.kt
@@ -17,7 +17,7 @@ class BeersRepository(
         remoteDataSource.fetch(collectionSize)
     }
 
-    fun like(id: Int) {
+    fun like(id: Long) {
         localDataSource.like(id)
     }
